\name{colorRampPC-package}
\alias{colorRampPC-package}
\docType{package}
\title{
    Provides pre-canned color-ramps, and functions for computing 'breaks' for plotting 
}
\description{
    Provides pre-canned color-ramps. 

    At the moment most color ramps from the QGIS 1-band-raster-colors plugin
    are included, along with the parts of cpt-city which are realeased under the
    GPL

    Also provides functions for computing 'breaks' for plotting, which also
    mimics some aspects of the raster colors plugin in QGIS. 

    Although designed for use with rasters/matrices etc, the methods are quite general
    and can be applied to any dataset with a min/max function, a quantile function, ... etc
}
\details{
\tabular{ll}{
Package: \tab colorRampPC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-15\cr
License: \tab GPL-3 \cr
}
}
\author{
Gareth Davies

Maintainer: Gareth Davies <grothered@gmail.com>
}
